+++ src/containers/Dashboard.js
@@ -2,13 +2,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
-import { Link, browserHistory } from 'react-router';
+import { browserHistory } from 'react-router';
 import * as actions from '../actions';
 
 import Balance from '../components/Balance';
 import Deposit from '../components/Deposit';
 import Withdrawl from '../components/Withdrawl';
-import Logout from '../components/Logout';
 
 class Dashboard extends Component {
   constructor(props) {
