+++ src/http/Client.js
@@ -1,42 +1,7 @@
-import _ from 'lodash';
+import Vault from './Vault';
 
-export const Client = function Client () {
-  var vault = function() {
-    var accounts = [
-      {
-        pin: '1234',
-        balance: '200.00'
-      },
-      {
-        pin: '1111',
-        balance: '10.06'
-      },
-      {
-        pin: '4444',
-        balance: '432.10'
-      }
-    ];
-
-    var vaultModule = {};
-    vaultModule.get = function(pin) { 
-      const index = _.findIndex(accounts, {pin});
-      if(index === undefined) {
-        return index;
-      }
-      return accounts[index];
-    }
-    vaultModule.put = function(pin, balance) { 
-      const index = _.findIndex(accounts, {pin});
-      if(index === undefined) {
-        return index;
-      }
-      const updatedAccount = { pin, balance };
-      accounts.splice(index, 1, updatedAccount);
-      return updatedAccount;
-    }
-    return vaultModule;
-  }
-
+export const Client = (function Client () {
+  var vault = new Vault();
   var teller = new vault();
 
   var clientModule = {};
@@ -52,7 +17,7 @@ export const Client = function Client () {
 
       case 'PUT':
         let newBalance = parseFloat(account.balance) + parseFloat(data.amount);
-        if(newBalance < account.balance) {
+        if(newBalance < 0) {
           Promise.reject({error: "insufficient-funds"});
         }
         const updatedAccount = teller.put(data.pin, newBalance);
@@ -64,4 +29,4 @@ export const Client = function Client () {
   }
 
   return clientModule;
-}
+})()
